# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campus Entity Resolution API",
    "description": "API for campus security monitoring and entity tracking",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/graph/timeline/{entity_id}": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Entity Timeline",
        "description": "Get chronological timeline for entity",
        "operationId": "get_entity_timeline_api_v1_graph_timeline__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/location/{location_id}/entities": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Entities At Location",
        "description": "Find entities at location at specific time",
        "operationId": "get_entities_at_location_api_v1_graph_location__location_id__entities_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Location Id" }
          },
          {
            "name": "timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Timestamp"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/alerts/missing": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Missing Entities",
        "description": "Find entities with no activity in last N hours",
        "operationId": "get_missing_entities_api_v1_graph_alerts_missing_get",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 72,
              "minimum": 1,
              "description": "Hours since last activity",
              "default": 12,
              "title": "Hours"
            },
            "description": "Hours since last activity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/graph/timeline/{entity_id}/summary": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Timeline Summary",
        "description": "Get human-readable timeline summary",
        "operationId": "get_timeline_summary_api_v1_graph_timeline__entity_id__summary_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/timeline/{entity_id}/with-gaps": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Timeline With Gaps",
        "description": "Get timeline with gap detection",
        "operationId": "get_timeline_with_gaps_api_v1_graph_timeline__entity_id__with_gaps_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "End Date"
            }
          },
          {
            "name": "gap_threshold_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "default": 2,
              "title": "Gap Threshold Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/timeline/{entity_id}/heatmap": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Activity Heatmap",
        "description": "Get activity heatmap data for visualization",
        "operationId": "get_activity_heatmap_api_v1_graph_timeline__entity_id__heatmap_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/timeline/{entity_id}/daily-summary": {
      "get": {
        "tags": ["graph"],
        "summary": "Get Daily Summary",
        "description": "Get detailed summary for a specific day",
        "operationId": "get_daily_summary_api_v1_graph_timeline__entity_id__daily_summary_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/timeline/{entity_id}/patterns": {
      "get": {
        "tags": ["graph"],
        "summary": "Detect Activity Patterns",
        "description": "Detect activity patterns and routines",
        "operationId": "detect_activity_patterns_api_v1_graph_timeline__entity_id__patterns_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/predict/location/{entity_id}": {
      "post": {
        "tags": ["graph"],
        "summary": "Predict Location",
        "description": "Predict entity location at target time using ML\n\nArgs:\n    entity_id: Entity to predict for\n    target_time: ISO timestamp to predict (default: now)\n    lookback_days: Days of history to use for prediction",
        "operationId": "predict_location_api_v1_graph_predict_location__entity_id__post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "target_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Target Time"
            }
          },
          {
            "name": "lookback_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Lookback Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/predict/gap/{entity_id}": {
      "post": {
        "tags": ["graph"],
        "summary": "Predict During Gap",
        "description": "Predict locations during an activity gap\n\nArgs:\n    entity_id: Entity ID\n    gap_start: Gap start time (ISO format)\n    gap_end: Gap end time (ISO format)",
        "operationId": "predict_during_gap_api_v1_graph_predict_gap__entity_id__post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entity Id" }
          },
          {
            "name": "gap_start",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Gap Start" }
          },
          {
            "name": "gap_end",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Gap End" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}